Implementing Upload and Download Files

+++ 21_WorkingWithFiles/rest-with-spring-boot-and-kotlin-erudio/src/test/kotlin/br/com/erudio/integrationtests/controller/withyml/BookControllerYamlTest.kt
@@ -288,7 +288,7 @@ class BookControllerYamlTest : AbstractIntegrationTest() {
     private fun mockBook() {
         book.title = "Docker Deep Dive"
         book.author = "Nigel Poulton"
-        book.price = (java.lang.Double.valueOf(55.99))
+        book.price = 55.99
         book.launchDate = Date()
     }
 }
